# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Radix Etymology API",
    "description": "Decode English vocabulary using Latin and Greek roots. Look up roots by form or meaning, and decompose academic words into their morphemes.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://radix-marketing.vercel.app"
    }
  ],
  "paths": {
    "/api/lookup": {
      "get": {
        "operationId": "lookupRoot",
        "summary": "Look up a root by form, meaning, or ID",
        "description": "Search for Latin/Greek roots by their form (e.g., 'bene', 'aud') or meaning (e.g., 'good', 'hear'). Returns matching roots with meanings, examples, and etymology.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query - a root form or meaning to search for",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "Look up a specific root by its ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "Language for translations (en, ko, vi)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ko", "vi"],
              "default": "en"
            }
          },
          {
            "name": "related",
            "in": "query",
            "description": "Include related roots (when using id parameter)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "default": "false"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with root information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Root not found"
          }
        }
      }
    },
    "/api/word": {
      "get": {
        "operationId": "decomposeWord",
        "summary": "Decompose an academic word into its morphemes",
        "description": "Break down an SAT/GRE vocabulary word into its Latin/Greek roots. Returns the morpheme breakdown with meanings and related words.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The word to decompose (e.g., 'beneficial', 'circumnavigate')",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search for words by prefix",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "Language for translations (en, ko, vi)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ko", "vi"],
              "default": "en"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Response format (json or ascii)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["json", "ascii"],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful word decomposition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Word not found in academic vocabulary"
          }
        }
      }
    }
  }
}